Add the new stdlib items from RFC 2700
Here we merely add the new items without touching any of the existing
definitions that would be deprecated by a full implementation of RFC 2700.

The next step would be to make the old items aliases for the new items
in order to maintain a single source of truth for these constants. We
deliberately refrain from taking that step in this PR; instead the goal
is to get these new unstable constants into the next beta so that we can
take that step without needing to deal with cfg to mollify the bootstrap.

Once the old items are defined in terms of the new items, the next step
would be to migrate rustc to using the new items.

Finally, the ultimate step would be for the old items to be deprecated
in the same commit that these new items are stabilized, pending a
stabilization issue.